Rewrote as a WPF application.
*****************
NEW FEATURES
*****************
The app now does not have to be relaunched or wait to attach/reattach to the TOS process.
There is a reset button to reset the xp per hour/time to level data.
The application position is saved between sessions.
The memory address is no longer a magic number harded coded in tosmonitor, it's now stored n the app.config along with the polling interval.