Fixed bug in experience control class where "Show" binding to the settings property was reversed.
Exposed the elapsed time in a new run time exp control that shows how long the current exp session has been.

Added an additional constructor argument to the experience control class for selecting the property that we want to listen for change events (this means we only update the Value when the property we care about changes).